made DataFetchingEnvironment an interface.  I appreciate that some one else did a PR on this but it had merge conflicts and I could not easily get the branch in play so it was easier to replicate the approach.
